+++ src/exit_free/exit.c
@@ -6,13 +6,13 @@
 /*   By: lvan-bre <lvan-bre@student.42lehavre.fr    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/07 20:02:50 by lvan-bre          #+#    #+#             */
-/*   Updated: 2025/05/20 19:09:50 by lvan-bre         ###   ########.fr       */
+/*   Updated: 2025/05/21 21:42:33 by lvan-bre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.functions.h"
 
-void	so_long_minishell(t_shell *data, int exitstat)
+void	so_long_minishell(t_shell *data, unsigned long exitstat)
 {
 	if (data->mlst)
 		freemlst(data->mlst);
@@ -28,7 +28,7 @@ void	so_long_minishell(t_shell *data, int exitstat)
 	exit(exitstat);
 }
 
-void	so_long_exec(t_shell *data, int exitstat, char *msg)
+void	so_long_exec(t_shell *data, unsigned long exitstat, char *msg)
 {
 	ft_freeall("%s%s%s%s%d%d%s%s", &data->line, &data->pwd, &data->old_pwd,
 		&data->home, &data->path, &data->envp, &msg, &data->prompt);
